+++ CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
 - Added notes to Firebase App Indexing.
 - Fixed issue that it scrolls to wrong position when changing translations.
 - Moved "search menu" to FAB.
+- Fixed continuous reading days while syncing.
 
 #### v1.16.3 (2016-12-26)
 - Fixed cursor jumping in notes due to synchronization.